feat(neovim): enable signature help, inlay hints and virtual lines
Signature help messages will be shown automatically, for example, in the
context of you calling a method it will show a floating window above
with the signature of that specific method, i.e.
`SomeClass.SomeMethod(string firstParameter)`.

Inlay hints are extra information that the LSP may provide; in C# for
example, methods and classes will have an inlay hint showing the amount
of references that particular method/class has.

Finally, virtual lines are exactly what they sound - fake lines added
onto your editor, which are used to display diagnostics. I've also added
the community recipe that makes so virtual lines are only shown when the
active line has a diagnostic.